Add simple random wanderer controller.
Sometimes walking, sometimes idleing. All random behaviour.
There is a crazy part in this code. The wanderer_random is a function which uses 'this'. And if you make this function the controller of a game object, 'this' is going to be the 'this' of that specific game object. And that is crazy in my opinion. Crazy but somehow cool. Making it possible to 'plug' controller functions like this into game objects, like stuffing a brain into a body.